Move statistics classes to separate package reduce dto sizes
