+++ src/main/java/com/tungstun/statistics/domain/statistics/util/CustomerStatisticsUtil.java
@@ -1,4 +1,4 @@
-package com.tungstun.statistics.domain.statistics;
+package com.tungstun.statistics.domain.statistics.util;
 
 import com.tungstun.barapi.domain.bill.Bill;
 import com.tungstun.barapi.domain.bill.Order;
