+++ src/main/java/com/tungstun/statistics/port/web/statistics/StatisticsController.java
@@ -5,9 +5,9 @@ import com.tungstun.statistics.application.statistics.StatisticsQueryHandler;
 import com.tungstun.statistics.application.statistics.query.GetBarStatistics;
 import com.tungstun.statistics.application.statistics.query.GetCustomerStatistics;
 import com.tungstun.statistics.application.statistics.query.GetGlobalCustomerStatistics;
-import com.tungstun.statistics.domain.statistics.BarStatistics;
-import com.tungstun.statistics.domain.statistics.CustomerStatistics;
-import com.tungstun.statistics.domain.statistics.GlobalCustomerStatistics;
+import com.tungstun.statistics.domain.statistics.model.BarStatistics;
+import com.tungstun.statistics.domain.statistics.model.CustomerStatistics;
+import com.tungstun.statistics.domain.statistics.model.GlobalCustomerStatistics;
 import com.tungstun.statistics.port.web.statistics.converter.BarStatisticsConverter;
 import com.tungstun.statistics.port.web.statistics.converter.CustomerStatisticsConverter;
 import com.tungstun.statistics.port.web.statistics.converter.GlobalCustomerStatisticsConverter;
