+++ src/main/java/com/tungstun/statistics/port/web/statistics/converter/BarStatisticsConverter.java
@@ -2,7 +2,7 @@ package com.tungstun.statistics.port.web.statistics.converter;
 
 import com.tungstun.barapi.port.web.bill.converter.BillConverter;
 import com.tungstun.barapi.port.web.order.converter.OrderProductConverter;
-import com.tungstun.statistics.domain.statistics.BarStatistics;
+import com.tungstun.statistics.domain.statistics.model.BarStatistics;
 import com.tungstun.statistics.port.web.statistics.response.BarStatisticsResponse;
 import org.springframework.stereotype.Component;
 
@@ -21,18 +21,10 @@ public class BarStatisticsConverter {
 
     public BarStatisticsResponse convert(BarStatistics barStatistics) {
         return new BarStatisticsResponse(
-                orderProductConverter.convert(barStatistics.mostSoldProductOfLastMonth()),
-                orderProductConverter.convert(barStatistics.mostSoldProductOfLastYear()),
-                orderProductConverter.convert(barStatistics.mostSoldProductOfAllTime()),
-                billConverter.convertToSummary(barStatistics.mostExpensiveBillOfLastMonth()),
-                billConverter.convertToSummary(barStatistics.mostExpensiveBillOfLastYear()),
-                billConverter.convertToSummary(barStatistics.mostExpensiveBillOfAllTime()),
-                barStatistics.totalSpentLastMonth(),
-                barStatistics.totalSpentOfLastYear(),
-                barStatistics.totalSpentOfAllTime(),
-                barStatistics.totalNotYetPayedLastMonth(),
-                barStatistics.totalNotYetPayedLastYear(),
-                barStatistics.totalNotYetPayedAllTime()
+                orderProductConverter.convert(barStatistics.mostSoldProduct()),
+                billConverter.convertToSummary(barStatistics.mostExpensiveBill()),
+                barStatistics.totalSpent(),
+                barStatistics.totalNotYetPayed()
         );
     }
