+++ src/main/java/com/tungstun/statistics/port/web/statistics/converter/GlobalCustomerStatisticsConverter.java
@@ -2,7 +2,7 @@ package com.tungstun.statistics.port.web.statistics.converter;
 
 import com.tungstun.barapi.port.web.bill.converter.BillConverter;
 import com.tungstun.barapi.port.web.order.converter.OrderProductConverter;
-import com.tungstun.statistics.domain.statistics.GlobalCustomerStatistics;
+import com.tungstun.statistics.domain.statistics.model.GlobalCustomerStatistics;
 import com.tungstun.statistics.port.web.statistics.response.GlobalCustomerStatisticsResponse;
 import org.springframework.stereotype.Component;
